pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plottk"
version = "0.1"
dependencies = [
  "pandas",
  "python-slugify",
  "tqdm",
  "wandb",
]
requires-python = ">=3.8"
authors = [
  {name = "Andrea Baisero"},
]
maintainers = [
  {name = "Andrea Baisero"},
]
description = "Plotting ToolKit!"
# readme = "README.rst"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = ["plotting"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python"
]

# [project.optional-dependencies]
# gui = ["PyQt5"]
# cli = [
#   "rich",
#   "click",
# ]

# [project.urls]
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
# Repository = "https://github.com/me/spam.git"
# "Bug Tracker" = "https://github.com/me/spam/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"


[tool.black]
skip_string_normalization = true
line_length = 80

[tool.isort]  # black compatibility
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
combine_as_imports = true
line_length = 80

[tool.ruff.format]
quote-style = 'single'

[tool.pylint.messages_control]  # black compatibility
disable = "C0103,C0114,C0115,C0116,C0330,C0326,R0801,R0902,R0903,R0911,R0913,R0914"
max_line_length = 120
